+++ apps/web/src/components/Stats/stackingdao/DepositsWithdrawBarChart/index.tsx
@@ -1,25 +1,11 @@
-import { db } from "@/db/db";
-import { transactionTable } from "@/db/schema";
-import { eq, sql } from "drizzle-orm";
+import { env } from "@/env";
+import type { StackingDAOProtocolStatsResponse } from "@/lib/api";
 import { DepositWithdrawBarChartClient } from "./DepositWithdrawBarChartClient";
 
-const getData = async () => {
-  const query = db
-    .select({
-      month: sql<string>`strftime('%Y-%m', timestamp, 'unixepoch') as month`,
-      depositsAmount: sql<number>`sum(case when action = 'stackingdao-deposit' then json->>'outAmount' else 0 end) as depositsAmount`,
-      withdrawalsAmount: sql<number>`sum(case when action = 'stackingdao-withdraw' then json->>'inAmount' else 0 end) as withdrawalsAmount`,
-    })
-    .from(transactionTable)
-    .where(eq(transactionTable.protocol, "stackingdao"))
-    .groupBy(sql`month`);
-
-  const stats = await query;
-  return stats;
-};
-
 export const DepositWithdrawBarChart = async () => {
-  const stats = await getData();
+  const stats: StackingDAOProtocolStatsResponse = await fetch(
+    `${env.NEXT_PUBLIC_API_URL}/api/protocols/stackingdao`,
+  ).then((res) => res.json());
 
   const formattedData: {
     date: string;
@@ -27,8 +13,8 @@ export const DepositWithdrawBarChart = async () => {
     deposits: number;
   }[] = stats.map((d) => ({
     date: d.month,
-    withdrawals: d.withdrawalsAmount,
-    deposits: d.depositsAmount,
+    withdrawals: d.withdrawals,
+    deposits: d.deposits,
   }));
 
   return <DepositWithdrawBarChartClient data={formattedData} />;