+++ apps/web/src/middleware.ts
@@ -3,8 +3,6 @@ import { env } from "./env";
 
 export const config = {
   matcher: [
-    // Protect the entire /api/chainhooks/* with a bearer token
-    "/api/chainhooks/:function*",
     // Protect the entire /api/cron/* with a bearer token
     "/api/cron/:function*",
   ],
@@ -25,14 +23,7 @@ export function middleware(request: NextRequest) {
   const authorization = request.headers.get("Authorization") || "";
   const [scheme, token] = authorization.split(" ");
 
-  if (endpoint[1] === "chainhooks") {
-    if (scheme !== "Bearer" || token !== env.CHAINHOOKS_API_TOKEN) {
-      return Response.json(
-        { success: false, message: "authentication failed" },
-        { status: 401 },
-      );
-    }
-  } else if (endpoint[1] === "cron") {
+  if (endpoint[1] === "cron") {
     if (scheme !== "Bearer" || token !== env.CRON_API_TOKEN) {
       return Response.json(
         { success: false, message: "authentication failed" },